+++ tests/common/Profile.cs
@@ -1,0 +1,14 @@
+
+namespace Xamarin.Tests
+{
+	public enum Profile
+	{
+		None,
+		iOS,
+		tvOS,
+		watchOS,
+		macOSMobile,
+		macOSFull,
+		macOSSystem,
+	}
+}
