+++ tests/mtouch/RegistrarTest.cs
@@ -9,6 +9,9 @@ using Xamarin.Tests;
 using Xamarin.Utils;
 using NUnit.Framework;
 
+using MTouchLinker = Xamarin.Tests.LinkerOption;
+using MTouchRegistrar = Xamarin.Tests.RegistrarOption;
+
 namespace Xamarin
 {
 	[TestFixture]
