+++ tests/mtouch/mtouch.csproj
@@ -57,6 +57,12 @@
       <Link>ProductTests.cs</Link>
     </Compile>
     <Compile Include="StringUtilsTest.cs" />
+    <Compile Include="..\common\Profile.cs">
+      <Link>Profile.cs</Link>
+    </Compile>
+    <Compile Include="..\common\BundlerTool.cs">
+      <Link>BundlerTool.cs</Link>
+    </Compile>
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
